perf(houston): threads for stats
